+++ database/migrations/2026_02_27_095754_create_role_permission_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('role_permission', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('role_permission')) {
+            Schema::create('role_permission', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('role_id')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->foreignId('role_id')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->foreignId('permission_id')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->foreignId('permission_id')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
     // RVERSE THE MIGRATION
     public function down(): void
